Design: A story of ampersand.
Adobe Caslon Pro, Italic variant offers a very nice and ornated ampersand.

The font Caslon Ampersand is a subset of it, limited to one glyph: the &.

You can print a beautiful ampersand with the following code:
<span class="ampersand">&</span>

[ Legal ]

This font should be recreated from a public domain element, and put in public
domain too.

Acceptable use. The font, restricted to one character, doesn't represent a work
protected by copyright, whereas (i) the character is a copy of an ampersand
designed by Robert Granjon, typograph in the second half of the sixteenth
century (ii) the cleaned shapes from this shape isn't subject to copyright per
Eltra Corp. v. Ringer (iii) the computer program to write this font is under the
shape of originality, becoming a trivial sample when used to render only a char.

[ Future ]

A new font, called "Granjon Ampersand", should be created, based directly on the
work of Robert Granjon.

This will produce the same shape, under public domain or a CC-0 / SIL / WTFPL.