Pluton: Document class. With XML description documents.
This class were started in www.dereckson.be.

This class detects if a document exists, and include it.

[ XML description documents]

It also reads document description (title, extra metadata) from a new
XML format, from _documents.xml file:

<documents topic="quux">
        <document>
                <article>one-tool</article>
                <title>One tool</title>
                <description>This is a tool.</description>
                <head><![CDATA[
  <link href='/stylesheets/forms.css'rel='stylesheet' type='text/css'>
                ]]></head>
        </document>
        <document>
                <article>other-tool</article>
                <title>Another tool</title>
        </document>
</documents>

This file belongs to the topic folder (here it will be quux/_documents.xml).

That XML idea will allow:
    - to get rid of the database for static sites;
    - to build a static generator.

The format is simple, so it won't be complicated to create an administration
panel to manage them.