New tools: Lists replace and operations
List replace allows to use PCRE regexp on every item of a list.
This is use to extract relevant info from a list and formats it.

List operations allows to add, substract or intersect lists.

This commit also fixes an issue on the Pluton URL handling
to play well with our lists URLs in Nginx.