Import Foundation Icon Fonts 3
Summary:
This uses a different CSS markup than version 1, so we can use both.

It's currently not possible to get rid of the former version,
as the following glyphs are missing from the new:

* tools (header)
* question (header)
* settings (footer)
* globe (footer)

The following icons are available in v3 but will stay v1 to be coherent
with the remaining:

* braille (header)
* flag (footer)

The v3 is intended to provide a richer UI at tool level.

So the offered guideline is v1 for the site layout, v3 for tools.

Test Plan:
Add a multi regexp mode to list/replace tool

This uses the `fi-page-multiple` icon.

Reviewers: dereckson

Differential Revision: http://devcentral.nasqueron.org/D108