list/replace: multi regexp support
Summary:
The fields expression and replacement are now handled a list of
expressions/replacements, delimited by new lines.

This allow to perform several regexp replacements in one operation.

This change affects only back-end code. A temptative UI change lives
at D107 but doesn't give satisfaction, as to change between a input
box (to get history) and a textarea (to encode seval expressions) is
rather tricky. A proper combobox/textarea control is needed. If the
UI uses React, it would also be easier to implement his, as we have
the state of the widget distinct of the representation.

Test Plan: live test at D107

Reviewers: dereckson

Maniphest Tasks: T626

Differential Revision: http://devcentral.nasqueron.org/D110