Fix 404 issues on nginx
Summary:
There is a kludge to get around a "known bug" in Document::find_document.

This kludge makes on a server not touched by this bug an attempt to reinclude /index.php instead to serve a 404 page.

To try to target every server configuration isn't the 2016 way.

This reverts a fix introduced in c35861b0.

Test Plan: Visit a 404 page on a nginx server

Reviewers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D715