Ignore empty PATH_INFO string
Summary:
If a webserver defines PATH_INFO, but with an empty string,
get_current_url() currently returns an empty string.

We now skip this detection method.

This isn't an issue not to trust "" or "0" values,
as we've better ways to detect this.

Test Plan:
Fire a Docker container with a php-fpm receiving
PATH_INFO through nginx to "".

Reviewers: dereckson

Reviewed By: dereckson

Tags: #keruald

Differential Revision: https://devcentral.nasqueron.org/D1165