Treat homepage like a regular page
Summary:
There is no strong rationale to treat differntly the homepage,
especially as we need to load _documents.xml configuration.

When there is a need to print the current URL, `get_current_url()`
can be used. Or to know if the homepage, `$document->is_homepage()`
can be called by the code needing that instead to define the homepage
as a no topic, no article document.

Test Plan: Tested on Dwellers

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1172