Switch gadgets/motd-variations to pure PHP
Summary:
This gadget was a proof of concept calling a TCL script,
but has been committed as is.

This change rewrites the TCL call in pure PHP.
It requires `jive` and `valspeak` as executables.

Reformat code to allow the HTML format to be prepended by spaces
and avoid too often to supersede 80 chars.

Ref T1294

Test Plan: Live at https://tools.nasqueron.org/gadgets/motd-variations

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1294

Differential Revision: https://devcentral.nasqueron.org/D1173